requirements.txt
streamlit
scikit-learn
numpy
joblib
torch
pykan
PyYAML
matplotlib
tqdm